refactor(data-model): move entity kind module
