+++ packages/shell/src/views/SchedulerGraphView.ts
@@ -1,6 +1,6 @@
 // Note: deno fmt can crash on deeply nested ternaries in html/svg templates - see renderBaselineStats
 import { css, html, LitElement, svg as svgTag, TemplateResult } from "lit";
-import { ENTITY_URI_SCHEMES } from "@commonfabric/data-model/fabric-primitives";
+import { ENTITY_URI_SCHEMES } from "@commonfabric/data-model/entity-kind";
 import { property, query, state } from "lit/decorators.js";
 import dagre from "dagre";
 import type { DebuggerController } from "../lib/debugger-controller.ts";
